refactor(mcp): reuse dcc adapter contract
