+++ crates/auroraview-mcp/README.md
@@ -107,81 +107,6 @@ Send a named event to the WebView's JavaScript context.
 }
 ```
 
-### `get_hwnd`
-
-Get the native window handle (HWND on Windows) for embedding in UE.
-
-**Parameters:**
-- `id` (optional): WebView ID
-
-**Returns:** `HwndOutput` with the handle value
-
-**Example:**
-```json
-{
-  "name": "get_hwnd",
-  "parameters": {}
-}
-```
-
-### `list_webviews`
-
-List all active WebView instances.
-
-**Returns:** `ListWebViewsOutput` with array of WebView info
-
-**Example:**
-```json
-{
-  "name": "list_webviews",
-  "parameters": {}
-}
-```
-
-### `create_webview`
-
-Create a new WebView instance.
-
-**Parameters:**
-- `title` (optional): Window title
-- `url` (optional): Initial URL
-- `html` (optional): Initial HTML
-- `width` (optional): Window width (default: 800)
-- `height` (optional): Window height (default: 600)
-- `visible` (optional): Whether window is visible (default: true)
-- `debug` (optional): Enable dev tools (default: false)
-
-**Returns:** `SuccessOutput` with new WebView ID
-
-**Example:**
-```json
-{
-  "name": "create_webview",
-  "parameters": {
-    "title": "My WebView",
-    "url": "https://example.com",
-    "width": 1024,
-    "height": 768
-  }
-}
-```
-
-### `close_webview`
-
-Close and remove a WebView instance.
-
-**Parameters:**
-- `id` (required): WebView ID to close
-
-**Returns:** `SuccessOutput`
-
-**Example:**
-```json
-{
-  "name": "close_webview",
-  "parameters": {"id": "550e8400-e29b-41d4-a716-446655440000"}
-}
-```
 
 ## AG-UI Protocol Support
